put clickableDonut in a separate file
